AIP-38 Integrate dag versions
