+++ airflow/ui/src/layouts/Details/DetailsLayout.tsx
@@ -16,11 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { Box, HStack, IconButton, ButtonGroup, Flex } from "@chakra-ui/react";
+import { Box, HStack, Flex } from "@chakra-ui/react";
 import { useReactFlow } from "@xyflow/react";
 import type { PropsWithChildren, ReactNode } from "react";
-import { FiGrid } from "react-icons/fi";
-import { MdOutlineAccountTree } from "react-icons/md";
 import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
 import { Outlet, useParams } from "react-router-dom";
 import { useLocalStorage } from "usehooks-ts";
@@ -37,6 +35,7 @@ import { OpenGroupsProvider } from "src/context/openGroups";
 import { DagBreadcrumb } from "./DagBreadcrumb";
 import { Graph } from "./Graph";
 import { Grid } from "./Grid";
+import { Header } from "./Header";
 import { NavTabs } from "./NavTabs";
 
 type Props = {
@@ -71,35 +70,8 @@ export const DetailsLayout = ({ children, error, isLoading, tabs }: Props) => {
       <Box flex={1} minH={0}>
         <PanelGroup autoSaveId={dagId} direction="horizontal">
           <Panel defaultSize={20} minSize={6}>
-            <Box height="100%" position="relative">
-              <ButtonGroup
-                attached
-                left={0}
-                position="absolute"
-                size="sm"
-                top={0}
-                variant="outline"
-                zIndex={1}
-              >
-                <IconButton
-                  aria-label="Show Grid"
-                  colorPalette="blue"
-                  onClick={() => setDagView("grid")}
-                  title="Show Grid"
-                  variant={dagView === "grid" ? "solid" : "outline"}
-                >
-                  <FiGrid />
-                </IconButton>
-                <IconButton
-                  aria-label="Show Graph"
-                  colorPalette="blue"
-                  onClick={() => setDagView("graph")}
-                  title="Show Graph"
-                  variant={dagView === "graph" ? "solid" : "outline"}
-                >
-                  <MdOutlineAccountTree />
-                </IconButton>
-              </ButtonGroup>
+            <Box height="100%" position="relative" pr={2}>
+              <Header dagId={dagId} dagView={dagView} setDagView={setDagView} />
               {dagView === "graph" ? <Graph /> : <Grid />}
             </Box>
           </Panel>
