+++ airflow/ui/src/layouts/Details/Graph/Graph.tsx
@@ -19,9 +19,10 @@
 import { useToken } from "@chakra-ui/react";
 import { ReactFlow, Controls, Background, MiniMap, type Node as ReactFlowNode } from "@xyflow/react";
 import "@xyflow/react/dist/style.css";
-import { useParams } from "react-router-dom";
+import { useParams, useSearchParams } from "react-router-dom";
 
 import { useGridServiceGridData, useStructureServiceStructureData } from "openapi/queries";
+import { SearchParamsKeys } from "src/constants/searchParams";
 import { useColorMode } from "src/context/colorMode";
 import { useOpenGroups } from "src/context/openGroups";
 import { isStatePending, useAutoRefresh } from "src/utils";
@@ -31,6 +32,8 @@ import { JoinNode } from "./JoinNode";
 import { TaskNode } from "./TaskNode";
 import type { CustomNodeProps } from "./reactflowUtils";
 import { useGraphLayout } from "./useGraphLayout";
+
+const VERSION_NUMBER_PARAM = SearchParamsKeys.VERSION_NUMBER;
 
 const nodeColor = (
   { data: { depth, height, isOpen, taskInstance, width }, type }: ReactFlowNode<CustomNodeProps>,
@@ -64,6 +67,9 @@ export const Graph = () => {
   const { colorMode = "light" } = useColorMode();
   const { dagId = "", runId, taskId } = useParams();
 
+  const [searchParams] = useSearchParams();
+  const selectedVersion = searchParams.get(VERSION_NUMBER_PARAM);
+
   // corresponds to the "bg", "bg.emphasized", "border.inverted" semantic tokens
   const [oddLight, oddDark, evenLight, evenDark, selectedDarkColor, selectedLightColor] = useToken("colors", [
     "white",
@@ -80,6 +86,7 @@ export const Graph = () => {
 
   const { data: graphData = { arrange: "LR", edges: [], nodes: [] } } = useStructureServiceStructureData({
     dagId,
+    versionNumber: selectedVersion === null ? undefined : parseInt(selectedVersion, 10),
   });
 
   const { data } = useGraphLayout({
